backend/requirements.txt
fastapi==0.82.0
pydantic[email]
tortoise-orm[asyncpg]
uvicorn
aerich
aiohttp
pymongo
